feat: Add Tour Details components and enhance TourDetailsPage
- Introduced TourContact, TourHeader, TourItinerary, and TourPricing components for better modularity.
- Updated TourDetailsPage to utilize new components, improving readability and maintainability.
- Enhanced itinerary data structure to include optional subtitles for each day.
- Implemented loading and not found states using LoadingSpinner and NotFoundTrek components.
- Refactored mobile and desktop layouts for better responsiveness and user experience.
