+++ frontend/src/components/index.ts
@@ -9,3 +9,6 @@ export { default as Footer } from './Footer';
 export { default as TrekCard } from './TrekCard';
 export { default as Breadcrumb } from './Breadcrumb';
 export { default as FilterDropdown } from './FilterDropdown';
+
+// Tour Details Components
+export * from './tour-details';
