+++ frontend/src/components/tour-details/index.ts
@@ -1,0 +1,14 @@
+export { default as TourBreadcrumb } from './TourBreadcrumb';
+export { default as TourCarousel } from './TourCarousel';
+export { default as TourBadges } from './TourBadges';
+export { default as TourHeader } from './TourHeader';
+export { default as TourPricing } from './TourPricing';
+export { default as TourContact } from './TourContact';
+export { default as TourAbout } from './TourAbout';
+export { default as TourItinerary } from './TourItinerary';
+export { default as ItineraryTimeline } from './ItineraryTimeline';
+export { default as ItineraryDayContent } from './ItineraryDayContent';
+export { default as LoadingSpinner } from './LoadingSpinner';
+export { default as NotFoundTrek } from './NotFoundTrek';
+export { default as MobileTourHeader } from './MobileTourHeader';
+export { default as MobileTourPricing } from './MobileTourPricing';
